code reference updates after file move
